PA/requirements.txt
pandas>=1.3.0
neo4j>=4.4.0
pyyaml>=6.0
python-dotenv>=0.19.0
numpy>=1.20.0
matplotlib>=3.4.0
seaborn>=0.11.0